init state

+++ controllers/contactsControllers.js
@@ -8,13 +8,15 @@ import {
 import * as contactsService from "../services/contactsServices.js";
 
 export const getAllContacts = async (req, res) => {
-  const data = await contactsService.listContacts();
+  const { id: owner } = req.user;
+  const data = await contactsService.listContacts({ owner });
   res.status(200).json(data);
 };
 
 export const getOneContact = async (req, res) => {
   const { id } = req.params;
-  const data = await contactsService.getContactById(id);
+  const { id: owner } = req.user;
+  const data = await contactsService.getContact({ id, owner });
 
   if (!data) {
     throw HttpError(404, "Not found");
@@ -24,7 +26,8 @@ export const getOneContact = async (req, res) => {
 
 export const deleteContact = async (req, res) => {
   const { id } = req.params;
-  const data = await contactsService.removeContact(id);
+  const { id: owner } = req.user;
+  const data = await contactsService.removeContact({ id, owner });
 
   if (!data) {
     throw HttpError(404, "Not found");
@@ -34,13 +37,15 @@ export const deleteContact = async (req, res) => {
 };
 
 export const createContact = async (req, res) => {
-  const data = await contactsService.addContact(req.body);
+  const { id: owner } = req.user;
+  const data = await contactsService.addContact({ ...req.body, owner });
   res.status(201).json(data);
 };
 
 export const updateContact = async (req, res) => {
   const { id } = req.params;
-  const result = await contactsService.updateContactByID(id, req.body);
+  const { id: owner } = req.user;
+  const result = await contactsService.updateContact({ id, owner }, req.body);
   res.json(result);
 };
